requirements.txt
pandas
jupyter 
pyarrow
ipykernel
sqlalchemy
psycopg2
dlt[duckdb]
dlt[filesystem]
dbt-core
dbt-postgres
dbt-duckdb
sqlfluff-templater-dbt
nbstripout